decimal digit accepted in withdraw and deposit

+++ src/bank/ATMNode.java
@@ -62,8 +62,8 @@ public class ATMNode extends RicartNode {
      * Check balance from LOCAL database
      */
     public String checkBalance(String user) {
-        int balance = localDB.getBalance(user);
-        if (balance == -1) {
+        double balance = localDB.getBalance(user);
+        if (balance == -1.0) {
             return "Error";
         }
         return String.valueOf(balance);
@@ -104,10 +104,10 @@ public class ATMNode extends RicartNode {
      */
     public String register(String user, String name, String pass, String amount) {
         try {
-            int initialBalance = Integer.parseInt(amount);
+            double initialBalance = Double.parseDouble(amount);
 
             // Validation
-            if (initialBalance < 500) {
+            if (initialBalance < 500.0) {
                 return "FAIL:MIN_DEPOSIT_500";
             }
 
@@ -177,7 +177,7 @@ public class ATMNode extends RicartNode {
                     String[] parts = response.split(":", 5);
                     if (parts.length >= 5) {
                         String name = parts[1];
-                        int balance = Integer.parseInt(parts[2]);
+                        double balance = Double.parseDouble(parts[2]);
                         String role = parts[3];
                         System.out.println("  ✅ Found account on Node " + peerId);
                         return new Database.Account(userId, name, password, balance, role);
@@ -276,7 +276,7 @@ public class ATMNode extends RicartNode {
                 String userId = parts[1];
                 String name = parts[2];
                 String password = parts[3];
-                int balance = Integer.parseInt(parts[4]);
+                double balance = Double.parseDouble(parts[4]);
                 String role = parts.length > 5 ? parts[5] : "user";
 
                 // Create account in local database (if doesn't exist)
@@ -290,7 +290,7 @@ public class ATMNode extends RicartNode {
             } else if ("REPLICATE_UPDATE".equals(action)) {
                 // REPLICATE_UPDATE:userId:newBalance
                 String userId = parts[1];
-                int newBalance = Integer.parseInt(parts[2]);
+                double newBalance = Double.parseDouble(parts[2]);
 
                 // Update balance in local database
                 if (localDB.accountExists(userId)) {
@@ -411,7 +411,7 @@ public class ATMNode extends RicartNode {
         for (String accStr : accounts) {
             String[] parts = accStr.split(":");
             if (parts.length >= 5) {
-                localDB.upsertAccount(parts[0], parts[1], parts[2], Integer.parseInt(parts[3]), parts[4]);
+                localDB.upsertAccount(parts[0], parts[1], parts[2], Double.parseDouble(parts[3]), parts[4]);
                 count++;
             }
         }
@@ -570,10 +570,10 @@ public class ATMNode extends RicartNode {
         try {
             // 1. DEPOSIT
             if ("DEPOSIT".equals(nextOperation)) {
-                int currentBalance = localDB.getBalance(opUser);
-                if (currentBalance != -1) {
-                    int amountObj = Integer.parseInt(opAmount);
-                    int newBalance = currentBalance + amountObj;
+                double currentBalance = localDB.getBalance(opUser);
+                if (currentBalance != -1.0) {
+                    double amountObj = Double.parseDouble(opAmount);
+                    double newBalance = currentBalance + amountObj;
                     localDB.updateBalance(opUser, newBalance);
 
                     // Broadcast replication
@@ -591,11 +591,11 @@ public class ATMNode extends RicartNode {
 
             // 2. WITHDRAW
             else if ("WITHDRAW".equals(nextOperation)) {
-                int currentBalance = localDB.getBalance(opUser);
-                if (currentBalance != -1) {
-                    int amountObj = Integer.parseInt(opAmount);
+                double currentBalance = localDB.getBalance(opUser);
+                if (currentBalance != -1.0) {
+                    double amountObj = Double.parseDouble(opAmount);
                     if (currentBalance >= amountObj) {
-                        int newBalance = currentBalance - amountObj;
+                        double newBalance = currentBalance - amountObj;
                         localDB.updateBalance(opUser, newBalance);
 
                         // Broadcast replication
@@ -616,14 +616,14 @@ public class ATMNode extends RicartNode {
 
             // 3. TRANSFER
             else if ("TRANSFER".equals(nextOperation)) {
-                int amountObj = Integer.parseInt(opAmount);
-                int senderBalance = localDB.getBalance(opUser);
-                int receiverBalance = localDB.getBalance(opTarget);
+                double amountObj = Double.parseDouble(opAmount);
+                double senderBalance = localDB.getBalance(opUser);
+                double receiverBalance = localDB.getBalance(opTarget);
 
                 // Validate sender has funds AND receiver exists
-                if (senderBalance != -1 && receiverBalance != -1 && senderBalance >= amountObj) {
-                    int newSenderBalance = senderBalance - amountObj;
-                    int newReceiverBalance = receiverBalance + amountObj;
+                if (senderBalance != -1.0 && receiverBalance != -1.0 && senderBalance >= amountObj) {
+                    double newSenderBalance = senderBalance - amountObj;
+                    double newReceiverBalance = receiverBalance + amountObj;
 
                     localDB.updateBalance(opUser, newSenderBalance);
                     localDB.updateBalance(opTarget, newReceiverBalance);
@@ -639,7 +639,7 @@ public class ATMNode extends RicartNode {
                     System.out.println("✅ ATM " + getNodeId() + ": Transferred $" + amountObj + " from " + opUser
                             + " to " + opTarget);
                 } else {
-                    if (receiverBalance == -1) {
+                    if (receiverBalance == -1.0) {
                         lastTransactionResult = "FAIL:RECEIVER_NOT_FOUND";
                     } else {
                         lastTransactionResult = "FAIL:INSUFFICIENT_FUNDS";
